refactor: extract EVM types to dedicated module
Move EVM-specific type definitions from src/types.ts to src/evm/types.ts for better code organization. Update imports across affected files while maintaining backward compatibility through re-exports.
